# swagger.yaml
openapi: 3.0.3
info:
  title: SET Service API
  version: 1.1.0
  description: |
    SET Service REST API for Worker, Company, Super Admin, and restricted Admin roles.

    Core worker flow:
    Worker registers -> verifies OTP -> becomes pending_approval -> super admin or permitted admin approves or rejects -> approved worker can log in and receive tokens.

    Contract-wide observability and authentication rules:
    - Every HTTP response includes `X-Request-ID` and `X-Correlation-ID`.
    - Clients may send either header using 8-128 characters from `[A-Za-z0-9._:-]`; invalid values are replaced server-side.
    - Error bodies include the effective `request_id`; the correlation identifier remains in the response header.
    - `BearerAuth` accepts access tokens only. Refresh tokens are valid only in the documented refresh/logout request bodies.
    - OTP request success means the request was accepted for delivery; it does not prove arrival at the handset or mailbox.
    - Interactive Swagger UI is disabled in production and should be exposed only in controlled non-production environments.
servers:
  - url: http://localhost:3000/v1
    description: Local development
  - url: https://api.setservice.az/v1
    description: Production
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT access token
      description: Short-lived SET Service JWT access token. Refresh tokens are rejected by authenticated API middleware.
...
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
...
  schemas:
...
    Error:
      type: object
      required: [error, code, request_id, timestamp]
      properties:
        error: { type: string, example: Worker is not approved for login. }
        code: { type: string, example: WORKER_NOT_APPROVED }
        details: { type: object, nullable: true, example: { status: pending_approval } }
        request_id: { $ref: '#/components/schemas/RequestId' }
        timestamp: { type: string, format: date-time }
...
    CreateKioskSessionRequest:
      type: object
      required: [assignment_id]
      properties:
        assignment_id: { type: string, format: uuid, example: "{{assignment_id}}" }
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: Optional kiosk display expiry. Defaults to server KIOSK_SESSION_TTL_HOURS.
    CreateVenueKioskRequest:
      type: object
      required: [name]
      properties:
        company_id: { type: string, format: uuid, example: "{{company_id}}" }
        name: { type: string, example: "Hilton əsas giriş" }
        location_label: { type: string, nullable: true, example: "Lobby / əsas giriş" }
    ActivateVenueKioskRequest:
      type: object
      required: [order_id]
      properties:
        order_id: { type: string, format: uuid, example: "{{order_id}}" }
        expires_at: { type: string, format: date-time, nullable: true }
...
    VenueKioskPublicContext:
      type: object
      properties:
        id: { type: string, format: uuid, example: "{{venue_kiosk_id}}" }
        kiosk_id: { type: string, format: uuid, example: "{{venue_kiosk_id}}" }
        company_id: { type: string, format: uuid, example: "{{company_id}}" }
        company_name: { type: string, example: "Hilton Baku" }
        name: { type: string, example: "Hilton əsas giriş" }
        kiosk_name: { type: string, example: "Hilton əsas giriş" }
        location_label: { type: string, nullable: true, example: "Lobby / əsas giriş" }
        status: { type: string, enum: [active, disabled], example: active }
        kiosk_status: { type: string, enum: [active, disabled], example: active }
        refresh_interval_seconds: { type: integer, example: 30 }
        active_session:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/VenueKioskActiveSession'
    VenueKiosk:
      allOf:
        - $ref: '#/components/schemas/VenueKioskPublicContext'
        - type: object
          properties:
            kiosk_url:
              type: string
              description: Authenticated management link. The capability is carried only in the URL fragment.
              example: "/kiosk#capability={{venue_kiosk_token}}"
    VenueKioskCreateResponse:
      allOf:
        - $ref: '#/components/schemas/VenueKiosk'
        - type: object
          properties:
            kiosk_token:
              type: string
              description: Raw persistent venue kiosk token. Returned on creation.
              example: "{{venue_kiosk_token}}"
...
    KioskQrTokenResponse:
      allOf:
        - $ref: '#/components/schemas/KioskSession'
        - type: object
          properties:
            token: { type: string, example: "{{attendance_qr_token}}" }
            expires_at: { type: string, format: date-time }
            refresh_after_seconds: { type: integer, example: 30 }
            kiosk_expires_at: { type: string, format: date-time, nullable: true }
...
paths:
...
  /attendance/venue-kiosks:
    get:
      tags: [Attendance]
      security: [{ BearerAuth: [] }]
      summary: List persistent venue kiosks
      parameters:
        - name: company_id
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Venue kiosks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/VenueKiosk' }
    post:
      tags: [Attendance]
      security: [{ BearerAuth: [] }]
      summary: Create persistent venue kiosk link
      description: Creates a stable venue/tablet kiosk URL. Admin later activates the current order/shift on this kiosk.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateVenueKioskRequest' }
      responses:
        '201':
          description: Venue kiosk created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VenueKioskCreateResponse' }
  /attendance/venue-kiosks/context:
    get:
      tags: [Attendance]
      summary: Get public persistent kiosk context
      description: |
        Capability-protected venue display context. The capability is sent only in
        X-Kiosk-Capability and is never echoed in the response or placed in an API path.
      parameters:
        - name: X-Kiosk-Capability
          in: header
          required: true
          schema: { type: string, minLength: 32 }
      responses:
        '200':
          description: Safe kiosk context
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VenueKioskPublicContext' }
        '404':
          description: Kiosk capability not found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '410':
          description: Kiosk disabled or revoked
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /attendance/venue-kiosks/qr-token:
    post:
      tags: [Attendance]
      summary: Generate fresh order-based 30-second venue kiosk QR
      description: The kiosk capability is accepted only in X-Kiosk-Capability and is never echoed.
      parameters:
        - name: X-Kiosk-Capability
          in: header
          required: true
          schema: { type: string, minLength: 32 }
      responses:
        '200':
          description: Fresh kiosk QR token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KioskQrTokenResponse' }
        '409':
          description: No active order selected for kiosk
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '410':
          description: Kiosk disabled/revoked or active order is no longer valid
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /attendance/venue-kiosks/{id}/activate:
    post:
      tags: [Attendance]
      security: [{ BearerAuth: [] }]
      summary: Activate an order/shift on a persistent venue kiosk
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActivateVenueKioskRequest' }
      responses:
        '200':
          description: Kiosk active session updated
          content: { application/json: { schema: { $ref: '#/components/schemas/VenueKiosk' } } }
  /attendance/venue-kiosks/{id}/active-session:
    delete:
      tags: [Attendance]
      security: [{ BearerAuth: [] }]
      summary: Deactivate the current order/shift on a venue kiosk
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Active session deactivated
          content: { application/json: { schema: { $ref: '#/components/schemas/VenueKiosk' } } }
  /attendance/venue-kiosks/{id}:
    delete:
      tags: [Attendance]
      security: [{ BearerAuth: [] }]
      summary: Disable a persistent venue kiosk
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Kiosk disabled
          content: { application/json: { schema: { $ref: '#/components/schemas/VenueKiosk' } } }
  /attendance/kiosk-sessions:
    post:
      tags: [Attendance]
      security: [{ BearerAuth: [] }]
      summary: Create standalone QR kiosk session
      description: |
        Creates a revocable, expirable kiosk display token for a partner venue tablet.
        The raw kiosk token is not an auth token and is returned only once.
        Requires super_admin or admin with manage_kiosks to create a kiosk for any accepted assignment. Approved company can create a kiosk only for its own accepted assignment.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateKioskSessionRequest' }
            examples:
              postman:
                value:
                  assignment_id: "{{assignment_id}}"
      responses:
        '201':
          description: Kiosk session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KioskSessionCreateResponse' }
              example:
                id: "{{kiosk_session_id}}"
                assignment_id: "{{assignment_id}}"
                order_id: "{{order_id}}"
                order_title: "Hilton Baku banket xidmeti"
                company_id: "{{company_id}}"
                company_name: "Hilton Baku"
                location: "Hilton Baku, Baki"
                category: "Ofisiant"
                status: active
                kiosk_status: active
                expires_at: "2030-01-15T22:00:00.000Z"
                revoked_at: null
                refresh_interval_seconds: 30
                kiosk_token: "{{kiosk_token}}"
                kiosk_url: "/kiosk#capability={{kiosk_token}}"
        '403':
          description: Company cannot create a kiosk for another company's assignment
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '404':
          description: Assignment not found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '409':
          description: Assignment is not accepted/active
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /attendance/kiosk-sessions/context:
    get:
      tags: [Attendance]
      summary: Get public QR kiosk display context
      description: |
        Capability-protected display context. X-Kiosk-Capability carries the
        capability; the response never returns it, auth credentials, or signing secrets.
      parameters:
        - name: X-Kiosk-Capability
          in: header
          required: true
          schema: { type: string, minLength: 32 }
      responses:
        '200':
          description: Safe kiosk context
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KioskSession' }
        '404':
          description: Kiosk capability not found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '410':
          description: Kiosk disabled, expired, revoked, or assignment/order inactive
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: "QR ekrani deaktiv edilib.", code: KIOSK_SESSION_INACTIVE }
  /attendance/kiosk-sessions/qr-token:
    post:
      tags: [Attendance]
      summary: Generate fresh 30-second kiosk QR token
      description: |
        X-Kiosk-Capability carries the kiosk capability outside the URL. Generates a fresh signed attendance QR token every 30 seconds.
        The previous displayed QR expires according to expires_at and must not be treated as valid after expiry.
      parameters:
        - name: X-Kiosk-Capability
          in: header
          required: true
          schema: { type: string, minLength: 32 }
      responses:
        '200':
          description: Fresh kiosk QR token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KioskQrTokenResponse' }
              example:
                id: "{{kiosk_session_id}}"
                assignment_id: "{{assignment_id}}"
                order_id: "{{order_id}}"
                order_title: "Hilton Baku banket xidmeti"
                company_id: "{{company_id}}"
                company_name: "Hilton Baku"
                token: "{{attendance_qr_token}}"
                expires_at: "2030-01-15T10:00:30.000Z"
                refresh_after_seconds: 30
        '404':
          description: Kiosk capability not found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '410':
          description: Kiosk disabled, expired, revoked, or assignment/order inactive
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /attendance/kiosk-sessions/{id}:
    delete:
      tags: [Attendance]
      security: [{ BearerAuth: [] }]
      summary: Revoke QR kiosk session
      description: Approved company can revoke its own kiosk sessions. Requires super_admin or admin with manage_kiosks to revoke any kiosk session.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Kiosk session revoked
        '403':
          description: Company cannot revoke another company's kiosk session
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '404':
          description: Kiosk session not found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /attendance/qr-token:
    post:
      tags: [Attendance]
      security: [{ BearerAuth: [] }]
      summary: Generate signed attendance QR token
      description: |
        Requires super_admin or admin with manage_kiosks to generate QR for any accepted assignment. Approved company can generate QR only for assignments on its own orders.
        The token is signed, assignment/order/company bound, and may be regenerated for dynamic display every 30 seconds.

        Postman setup:
        1. Create an assignment and have the worker accept it.
        2. Save the assignment id as {{assignment_id}}.
        3. Call this endpoint and save token as {{attendance_qr_token}}.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerateAttendanceQrRequest' }
            examples:
              postman:
                value:
                  assignment_id: "{{assignment_id}}"
                  ttl_seconds: 60
      responses:
        '200':
          description: QR token generated
          content:
            application/json:
              example:
                assignment_id: "{{assignment_id}}"
                order_id: "{{order_id}}"
                order_title: "Hotel banquet service shift"
                company_id: "{{company_id}}"
                company_name: "SET Service MMC"
                token: "{{attendance_qr_token}}"
                expires_at: "2030-01-15T10:05:00.000Z"
                refresh_after_seconds: 30
        '403':
          description: Company cannot generate QR for this assignment
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '404':
          description: Assignment not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                assignmentNotFound:
                  value:
                    error: Assignment not found.
                    code: ASSIGNMENT_NOT_FOUND
        '409':
          description: Assignment is not accepted/active
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                assignmentNotAccepted:
                  value:
                    error: Assignment must be accepted and active before QR generation.
                    code: ASSIGNMENT_NOT_ACCEPTED
  /attendance/check-in:
    post:
      tags: [Attendance]
      security: [{ BearerAuth: [] }]
      summary: Check in to own accepted assignment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttendanceCheckInRequest' }
            examples:
              postman:
                value:
                  assignment_id: "{{assignment_id}}"
                  qr_token: "{{attendance_qr_token}}"
                  location:
                    latitude: 40.4093
                    longitude: 49.8671
                    address: Grand Hotel Baku entrance
                  notes: Arrived on time.
      responses:
        '201':
          description: Check-in recorded
          content: { application/json: { schema: { $ref: '#/components/schemas/AttendanceLog' } } }
        '400':
          description: Validation error
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '401':
          description: QR token invalid
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                invalid:
                  value: { error: QR token is invalid., code: QR_TOKEN_INVALID }
        '403':
          description: Worker cannot check in for another worker assignment
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '404':
          description: Assignment not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                assignmentNotFound:
                  value: { error: Assignment not found., code: ASSIGNMENT_NOT_FOUND }
        '409':
          description: Assignment not accepted, duplicate open attendance, or completed attendance exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                notAccepted:
                  value: { error: Assignment must be accepted and active before check-in., code: ASSIGNMENT_NOT_ACCEPTED }
                alreadyCheckedIn:
                  value: { error: Attendance is already checked in for this assignment., code: ATTENDANCE_ALREADY_CHECKED_IN }
                alreadyCompleted:
                  value: { error: Attendance session is already completed for this assignment., code: ATTENDANCE_ALREADY_COMPLETED }
        '410':
          description: QR token expired
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                expired:
                  value: { error: QR token has expired., code: QR_TOKEN_EXPIRED }
  /attendance/check-out:
    post:
      tags: [Attendance]
      security: [{ BearerAuth: [] }]
      summary: Check out from own accepted assignment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttendanceCheckOutRequest' }
            examples:
              postman:
                value:
                  assignment_id: "{{assignment_id}}"
                  qr_token: "{{attendance_qr_token}}"
                  notes: Shift completed.
      responses:
        '200':
          description: Check-out recorded
          content: { application/json: { schema: { $ref: '#/components/schemas/AttendanceLog' } } }
        '400':
          description: No open check-in exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                notCheckedIn:
                  value: { error: Attendance has no open check-in for this assignment., code: ATTENDANCE_NOT_CHECKED_IN }
        '401':
          description: QR token invalid
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                invalid:
                  value: { error: QR token is invalid., code: QR_TOKEN_INVALID }
        '403':
          description: Worker cannot check out another worker assignment
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '404':
          description: Assignment not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                assignmentNotFound:
                  value: { error: Assignment not found., code: ASSIGNMENT_NOT_FOUND }
        '409':
          description: Assignment is not accepted/active
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                notAccepted:
                  value: { error: Assignment must be accepted and active before checkout., code: ASSIGNMENT_NOT_ACCEPTED }
        '410':
          description: QR token expired
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                expired:
                  value: { error: QR token has expired., code: QR_TOKEN_EXPIRED }
  /attendance:
    get:
      tags: [Attendance]
      security: [{ BearerAuth: [] }]
      summary: List attendance records
      description: Requires super_admin or admin with view_attendance to see all. Company sees own order assignment attendance. Worker sees own attendance.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: assignment_id
          in: query
          schema: { type: string, format: uuid }
        - name: order_id
          in: query
          schema: { type: string, format: uuid }
        - name: worker_id
          in: query
          schema: { type: string, format: uuid }
        - name: open_only
          in: query
          schema: { type: boolean }
        - name: sort
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        '200':
          description: Paginated attendance records
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttendanceListResponse' }
              example:
                data:
                  - id: "{{attendance_id}}"
                    assignment_id: "{{assignment_id}}"
                    checkin_time: "2030-01-15T10:00:00.000Z"
                    checkout_time: "2030-01-15T18:00:00.000Z"
                    duration_minutes: 480
                    checkin_location:
                      latitude: 40.4093
                      longitude: 49.8671
                      address: Grand Hotel Baku entrance
                    checkout_location:
                      latitude: 40.4093
                      longitude: 49.8671
                      address: Grand Hotel Baku exit
                    checkin_notes: Arrived on time.
                    checkout_notes: Shift completed.
                    created_at: "2030-01-15T10:00:00.000Z"
                    updated_at: "2030-01-15T18:00:00.000Z"
                    assignment:
                      id: "{{assignment_id}}"
                      status: accepted
                      worker:
                        id: "{{worker_id_1}}"
                        name: Ali Hasanov
                        phone: "+994501234567"
                      order:
                        id: "{{order_id}}"
                        title: Banquet service shift
                        status: active
                        company:
                          id: "{{company_id}}"
                          name: Grand Hotel Baku
                meta:
                  page: 1
                  limit: 20
                  total: 1
                  total_pages: 1
        '403':
          description: Account cannot view requested attendance records
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                forbidden:
                  value: { error: Workers can view only their own attendance records., code: FORBIDDEN }
  /attendance/{id}:
    get:
      tags: [Attendance]
      security: [{ BearerAuth: [] }]
      summary: Get attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Attendance record
          content: { application/json: { schema: { $ref: '#/components/schemas/AttendanceLog' } } }
        '403':
          description: Account cannot view this attendance record
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '404':
          description: Attendance not found or not visible
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...